move errors to models package
