+++ sdk/core/azure-core/src/main/java/com/azure/core/exception/HttpResponseException.java
@@ -4,6 +4,7 @@
 package com.azure.core.exception;
 
 import com.azure.core.http.HttpResponse;
+import com.azure.core.models.HttpResponseError;
 
 /**
  * The exception thrown when an unsuccessful response is received with http status code (e.g. 3XX, 4XX, 5XX) from the
