--- sdk/core/azure-core/src/main/java/com/azure/core/implementation/exception/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Package containing implementation-specific exception APIs that should not be used by end-users.
- */
-package com.azure.core.implementation.exception;
