+++ sdk/core/azure-core/src/main/java/com/azure/core/models/HttpResponseInnerError.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.exception;
+package com.azure.core.models;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
